+++ test2/amount-duration-conversion-data.ts
